7/11/2020- 7/14/2020 Programmed new localization system. It is essentially a copy of last year localization but in kotlin
7/16
Set up a PID class and Mecanum controller that uses the PID to hold rotation.

7/27/2020
Updated the PID controller class that takes into account time. Previously, the derivative and integral would take the same action each step regardless of how much time there was in between steps. This would lead to fluctuating power based on fluctuating program speed. Now it adjusts the integral and derivative so that power doesn't fluctuate based on program speed. The effects of this change were immediately noticeable in the performance of the pid controller. There was no persistent oscillation. We suspect the reason for this is that when the power was fluctuating there would occasionally power spikes that would cause the robot to overshoot the target.

7/28/2020
Implemented new Point to point controller. The new point to point controller had a couple key differences from last years. The first difference is that is uses the new PID controller that takes into account time. This alone greatly improved performance as it would travel in a much more stable manner. The next big difference was the way rotation is handled. The robot is constantly trying to adjust its rotation to the goal using a PID as opposed to only adjusting the rotation at the end or using a proportional controller while moving and a PID at the end. This way the robot doesn't have to switch between movement and rotation while settling. Finally, we are using velocity as a goal system instead of distance. Instead of moving onto the next task when the robot is close enough to the goal, it moves on once it has gotten below a certain velocity. This prevents the robot from sitting in one place stalled for countless seconds.

A bug was discovered in the localizer while developing the point to point controller. The rotation convention (clockwise/anti-clockwise) was flipped so that rotating the robot 90 degrees would yield the opposite of the direction vector to goal. This was fixed by making the rotation convention through out the program consistently anti-clockwise

7/30/2020
Made some minor changes and bug fixes. The main change made was that the the robots rotation had to be within a certain distance of the goal to avoid it stopping with a rotation that is inaccurate.
A severe bug in the point to point controller was also discovered. The goal rotation was being set to the goal minus the robot rotation rather than just the goal rotation. This was due to the fact we were determining the direction to go by transforming the origin of the goal point to the robots position. However this also affected the goal rotation which we forgot to account for. To fix this we overrode the adjusted rotation with the true goal rotation.